+++ opentelemetry-sdk/tests/trace/propagation/test_b3_format.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 import unittest
-from unittest.mock import patch
+from unittest.mock import Mock, patch
 
 import opentelemetry.sdk.trace as trace
 import opentelemetry.sdk.trace.propagation.b3_format as b3_format
@@ -65,6 +65,14 @@ class TestB3Format(unittest.TestCase):
         cls.serialized_parent_id = b3_format.format_span_id(
             ids_generator.generate_span_id()
         )
+
+    def setUp(self) -> None:
+        tracer_provider = trace.TracerProvider()
+        patcher = unittest.mock.patch.object(
+            trace_api, "get_tracer_provider", return_value=tracer_provider
+        )
+        patcher.start()
+        self.addCleanup(patcher.stop)
 
     def test_extract_multi_header(self):
         """Test the extraction of B3 headers."""
@@ -329,3 +337,21 @@ class TestB3Format(unittest.TestCase):
 
         ctx = FORMAT.extract(CarrierGetter(), {})
         FORMAT.inject(setter, {}, ctx)
+
+    def test_fields(self):
+        """Make sure the fields attribute returns the fields used in inject"""
+
+        tracer = trace.TracerProvider().get_tracer("sdk_tracer_provider")
+
+        mock_set_in_carrier = Mock()
+
+        with tracer.start_as_current_span("parent"):
+            with tracer.start_as_current_span("child"):
+                FORMAT.inject(mock_set_in_carrier, {})
+
+        inject_fields = set()
+
+        for call in mock_set_in_carrier.mock_calls:
+            inject_fields.add(call[1][1])
+
+        self.assertEqual(FORMAT.fields, inject_fields)