+++ opentelemetry-sdk/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Add meter reference to observers
   ([#1425](https://github.com/open-telemetry/opentelemetry-python/pull/1425))
+- Add `fields` to propagators
+  ([#1374](https://github.com/open-telemetry/opentelemetry-python/pull/1374))
 
 ## Version 0.16b0